# Conflicts:
#	.idea/compiler.xml
#	.idea/misc.xml
#	Style Guide.txt
#	src/main/java/IO_Package/IO_Manager.java
#	src/main/java/Instances/Agents/Agent.java
#	src/main/java/Instances/InstanceBuilder_BGU.java
#	src/main/java/Instances/InstanceManager.java
#	src/main/java/Instances/MAPF_Instance.java
#	src/main/java/Instances/Maps/Coordinate_2D.java
#	src/main/java/Instances/Maps/Enum_MapCellType.java
#	src/main/java/Instances/Maps/GraphMap.java
#	src/main/java/Instances/Maps/GraphMapCell.java
#	src/main/java/Instances/Maps/I_Map.java
#	src/main/java/Instances/Maps/I_MapCell.java
#	src/main/java/Instances/Maps/MapFactory.java
#	src/main/java/Metrics/InstanceReport.java
#	src/main/java/Metrics/S_Metrics.java

+++ src/main/java/Instances/Maps/I_Coordinate.java
@@ -1,7 +1,25 @@
 package Instances.Maps;
 
+/**
+ * Represents a unique location in Euclidean space.
+ */
 public interface I_Coordinate {
-
+    // todo generalize to include non-euclidean space? replace methods with one abstract h() function?
+    /**
+     * Returns the euclidean distance to another {@link I_Coordinate coordinate}. Should return 0 iff this.equals(other)
+     * return true.  If other is null, or is not of the same runtime type as this, returns -1.
+     * @param other a {@link I_Coordinate coordinate}.
+     * @return the euclidean distance to another {@link I_Coordinate coordinate}. If other is null, or is not of the
+     * same runtime type as this, returns -1.
+     */
     float euclideanDistance(I_Coordinate other);
-    float manhattanDistance(I_Coordinate other);
+
+    /**
+     * Returns the manhattan distance to another {@link I_Coordinate coordinate}. Should return 0 iff this.equals(other)
+     * return true.  If other is null, or is not of the same runtime type as this, returns -1.
+     * @param other a {@link I_Coordinate coordinate}.
+     * @return the manhattan distance to another {@link I_Coordinate coordinate}.  If other is null, or is not of the
+     * same runtime type as this, returns -1.
+     */
+    int manhattanDistance(I_Coordinate other);
 }
